requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-rest-knox==4.1.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.3
gunicorn==19.9.0
pep8==1.7.1
Pillow==6.2.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
